+++ semcore/checkbox/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [7.21.1] - 2024-02-05
+## [7.21.2] - 2024-02-06
 
 ### Fixed
 
 - Invalid attribute errors in console that appears when component has invalid state.
+
+## [7.21.1] - 2024-02-06
+
+### Fixed
+
+- `Checkbox.Text` custom color were not working.
 
 ## [7.21.0] - 2024-02-01